+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Usage:
 ### 1.17-SNAPSHOT
 * Fix #2335: Add support for configuring nodeSelector spec for controller via xml/groovy DSL configuration
 * Fix #2459: Allow configuring Buildpacks build via ImageConfiguration
+* Fix #2860: Correctly pass Docker build-arg from the build configuration to the Openshift build strategy
 
 ### 1.16.2 (2024-03-27)
 * Fix #2461: `k8s:watch`/`k8sWatch` should throw error in `buildpacks` build strategy